requirements.txt
asgiref==3.3.1
boto3==1.17.5
botocore==1.20.5
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.1.5
django-cache-url==3.2.3
django-crispy-forms==1.10.0
django-on-heroku==1.1.2
django-storages==1.11.1
environs==8.0.0
gunicorn==20.0.4
heroku==0.1.1
idna==2.10
jmespath==0.10.0
marshmallow==3.10.0
Pillow==8.1.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.3
whitenoise==5.2.0
